adding update for the stucture file
split the main file to another file and problem in handleClickCard
