+++ index.html
@@ -13,6 +13,7 @@
   <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css" />
   <!-- Axios import -->
   <script src="https://cdn.jsdelivr.net/npm/axios/dist/axios.min.js"></script>
+    <script src="https://cdn.jsdelivr.net/npm/axios/dist/axios.min.js"></script>
   <style>
     body {
       background-color: #171516f2;
@@ -438,7 +439,7 @@
     <!--// adding loading for the pagination // -->
   </div>
   <!-- // start post //  -->
-  <script src="script.js"></script>
+  <script type="module" src="script.mjs"></script>
 </body>
 
 </html>
