+++ package.json
@@ -12,5 +12,6 @@
   "dependencies": {
     "axios": "^1.6.7",
     "bootstrap": "^5.3.2"
-  }
+  },
+  "type": "module"
 }
