+++ src/js/index.mjs
@@ -1,0 +1,9 @@
+import reloadWindow from "./UI/reloadWindow.mjs";
+import loaderHandler from "./UI/loaderHandler.mjs";
+import setLocalStorageInfo from "./Logic/localStorage.mjs";
+import setupUi from "./UI/setupUi.mjs";
+import createAlert from "./UI/createAlert.mjs";
+import createUserLoginInfoNavBar from "./UI/createUserInfoNav.mjs";
+import { templateCard } from "./templates/templateCard.mjs";
+
+export { reloadWindow, loaderHandler, setLocalStorageInfo, setupUi, createAlert , createUserLoginInfoNavBar , templateCard};
